Dashboard-main/requirements.txt
# before running project, run pip install -r requirements.txt to install all requirements
# please add any new requirements here
blinker==1.9.0
click==8.1.8
Faker==37.1.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
